// frontend/src/shared/lib/api/model/loanItemStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Cellar
 * OpenAPI spec version: 0.1.0
 */

/**
 * Status of a plate within a loan.
 */
export type LoanItemStatus = typeof LoanItemStatus[keyof typeof LoanItemStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const LoanItemStatus = {
  requested: 'requested',
  approved: 'approved',
  checked_out: 'checked_out',
  return_pending: 'return_pending',
  returned: 'returned',
  denied: 'denied',
  cancelled: 'cancelled',
} as const;