// frontend/src/shared/lib/api/model/loanResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Cellar
 * OpenAPI spec version: 0.1.0
 */
import type { LoanResponseApprovedBy } from './loanResponseApprovedBy';
import type { LoanResponseDueDate } from './loanResponseDueDate';
import type { LoanResponseNotes } from './loanResponseNotes';
import type { LoanStatus } from './loanStatus';
import type { LoanResponseClosedAt } from './loanResponseClosedAt';
import type { LoanItemResponse } from './loanItemResponse';

export interface LoanResponse {
  id: string;
  workspace_id: string;
  owner_org_id: string;
  borrower_org_id: string;
  requested_by: string;
  approved_by?: LoanResponseApprovedBy;
  due_date?: LoanResponseDueDate;
  notes?: LoanResponseNotes;
  status: LoanStatus;
  closed_at?: LoanResponseClosedAt;
  created_at: string;
  items: LoanItemResponse[];
  version: number;
}