// frontend/src/shared/lib/api/model/loanStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Cellar
 * OpenAPI spec version: 0.1.0
 */

/**
 * Lifecycle status of a plate loan.
 */
export type LoanStatus = typeof LoanStatus[keyof typeof LoanStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const LoanStatus = {
  open: 'open',
  closed: 'closed',
} as const;